:truck: Move AntragListSearch to antragList directory
